dbeaver/pro#7352 remove cli dependency from model bundles
