+++ server/bundles/io.cloudbeaver.model.cli/META-INF/MANIFEST.MF
@@ -1,0 +1,13 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Vendor: DBeaver Corp
+Bundle-Name: Cloudbeaver Web Model
+Bundle-SymbolicName: io.cloudbeaver.model.cli;singleton:=true
+Bundle-Version: 1.0.0.qualifier
+Bundle-Release-Date: 20251222
+Bundle-RequiredExecutionEnvironment: JavaSE-21
+Bundle-ActivationPolicy: lazy
+Require-Bundle: org.jkiss.dbeaver.model.cli;visibility:=reexport
+Export-Package: io.cloudbeaver.model.cli
+Automatic-Module-Name: io.cloudbeaver.model.cli
+Import-Package: org.slf4j
