+++ server/bundles/io.cloudbeaver.model.cli/build.properties
@@ -1,0 +1,5 @@
+source.. = src/
+output.. = target/classes/
+bin.includes = .,\
+    META-INF/,\
+    plugin.xml
