+++ server/bundles/io.cloudbeaver.model.cli/plugin.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.2"?>
+
+<plugin>
+
+</plugin>
