+++ server/bundles/io.cloudbeaver.model.cli/src/io/cloudbeaver/model/cli/CloudbeaverCliConstants.java
