+++ server/bundles/io.cloudbeaver.model/META-INF/MANIFEST.MF
@@ -9,7 +9,6 @@ Bundle-RequiredExecutionEnvironment: JavaSE-21
 Bundle-ActivationPolicy: lazy
 Require-Bundle: org.jkiss.dbeaver.data.gis;visibility:=reexport,
  org.jkiss.dbeaver.model;visibility:=reexport,
- org.jkiss.dbeaver.model.cli;visibility:=reexport,
  org.jkiss.dbeaver.model.sm;visibility:=reexport,
  org.jkiss.dbeaver.model.event;visibility:=reexport,
  org.jkiss.dbeaver.model.nio;visibility:=reexport,
@@ -30,13 +29,10 @@ Export-Package: io.cloudbeaver,
  io.cloudbeaver.model.app,
  io.cloudbeaver.model.config,
  io.cloudbeaver.model.apilog,
- io.cloudbeaver.model.cli,
- io.cloudbeaver.model.events,
  io.cloudbeaver.model.fs,
  io.cloudbeaver.model.log,
  io.cloudbeaver.model.rm,
  io.cloudbeaver.model.rm.local,
- io.cloudbeaver.model.rm.lock,
  io.cloudbeaver.model.session,
  io.cloudbeaver.model.user,
  io.cloudbeaver.model.utils,
