+++ server/bundles/io.cloudbeaver.model/src/io/cloudbeaver/model/app/BaseServletApplication.java
@@ -16,7 +16,6 @@
  */
 package io.cloudbeaver.model.app;
 
-import io.cloudbeaver.model.cli.CloudBeaverInstanceServer;
 import io.cloudbeaver.model.log.SLF4JLogHandler;
 import org.eclipse.core.runtime.Platform;
 import org.eclipse.equinox.app.IApplicationContext;
@@ -28,7 +27,6 @@ import org.jkiss.dbeaver.model.DBFileController;
 import org.jkiss.dbeaver.model.app.DBPWorkspace;
 import org.jkiss.dbeaver.model.auth.SMCredentialsProvider;
 import org.jkiss.dbeaver.model.auth.SMSessionContext;
-import org.jkiss.dbeaver.model.cli.ApplicationInstanceController;
 import org.jkiss.dbeaver.model.data.json.JSONUtils;
 import org.jkiss.dbeaver.model.impl.app.ApplicationRegistry;
 import org.jkiss.dbeaver.model.impl.app.BaseApplicationImpl;
@@ -41,7 +39,6 @@ import org.jkiss.dbeaver.utils.GeneralUtils;
 import org.jkiss.dbeaver.utils.RuntimeUtils;
 import org.jkiss.utils.CommonUtils;
 
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
@@ -60,7 +57,6 @@ public abstract class BaseServletApplication extends BaseApplicationImpl impleme
     private static final Log log = Log.getLog(BaseServletApplication.class);
 
     private String instanceId;
-    private CloudBeaverInstanceServer instanceServer;
     @Override
     public RMController createResourceController(
         @NotNull SMCredentialsProvider credentialsProvider,
@@ -204,21 +200,12 @@ public abstract class BaseServletApplication extends BaseApplicationImpl impleme
     public Object start(IApplicationContext context) {
         initializeApplicationServices();
         try {
-            try {
-                this.instanceServer = createInstanceServer();
-            } catch (Exception e) {
-                log.error("Error initializing instance server", e);
-            }
             startServer();
         } catch (Exception e) {
             log.error(e.getMessage(), e);
             return EXIT_ERROR_UNSPECIFIED;
         }
         return EXIT_OK;
-    }
-
-    protected CloudBeaverInstanceServer createInstanceServer() throws IOException {
-        return new CloudBeaverInstanceServer();
     }
 
     protected abstract void startServer() throws DBException;
@@ -279,12 +266,6 @@ public abstract class BaseServletApplication extends BaseApplicationImpl impleme
         } catch (Exception e) {
             log.error("Failed close " + name, e);
         }
-    }
-
-    @Nullable
-    @Override
-    public ApplicationInstanceController getInstanceServer() {
-        return instanceServer;
     }
 
     @Override
