+++ server/bundles/io.cloudbeaver.server/META-INF/MANIFEST.MF
@@ -18,6 +18,7 @@ Require-Bundle: org.eclipse.core.runtime;visibility:=reexport,
  org.jkiss.bundle.apache.dbcp,
  org.jkiss.dbeaver.net.ssh,
  io.cloudbeaver.model;visibility:=reexport,
+ io.cloudbeaver.model.cli,
  io.cloudbeaver.service.security;visibility:=reexport
 Export-Package: io.cloudbeaver,
  io.cloudbeaver.model,
