+++ server/bundles/pom.xml
@@ -14,6 +14,7 @@
 
     <modules>
         <module>io.cloudbeaver.model</module>
+        <module>io.cloudbeaver.model.cli</module>
         <module>io.cloudbeaver.server</module>
         <module>io.cloudbeaver.server.ce</module>
         <module>io.cloudbeaver.slf4j</module>
